Separate search service class, separate RpslObjectStreamer class.
